refactor: remove old Twig extensions no longer required or getting used; refactor Color logic into standalone Class; wire up last of the Image-related Twig extensions
