+++ packages/core-php/src/TwigExtensions/BoltExtras.php
@@ -15,6 +15,7 @@ class BoltExtras extends Twig_Extension implements Twig_ExtensionInterface {
       Bolt\TwigFunctions::deep_merge(),
       Bolt\TwigFunctions::color_contrast(),
       Bolt\TwigFunctions::create_attribute(),
+      Bolt\TwigFunctions::link(),
     ];
   }
