+++ packages/core-php/src/Utils.php
@@ -10,4 +10,14 @@ class Utils {
     return MarkdownExtra::defaultTransform($string);
   }
 
+  // Given a relative file path, return the resolved absolute path
+  public static function get_absolute_path($relativeFilePath) {
+    $publicDir = '/dist';
+    return getcwd() . $publicDir . $relativeFilePath;
+  }
+
+  // Given an absolute file path, return the file's extension info
+  public static function get_file_ext($absoluteFilePath) {
+    return pathinfo($absoluteFilePath, PATHINFO_EXTENSION);
+  }
 }
