--- packages/twig-extensions/filters/custom-filters/complimentary_color.filter.php
@@ -1,11 +1,0 @@
-<?php
-
-use Mexitek\PHPColors\Color;
-
-$filter = new Twig_SimpleFilter('complimentary_color', function ($color) {
-  
-  $myColor = new Color($color);
-  
-  return $myColor->complementary();
-
-});
