--- packages/twig-extensions/filters/drupal-filters/clean_class.filter.php
@@ -1,5 +1,0 @@
-<?php
-
-$filter = new Twig_SimpleFilter('clean_class', function ($string) {
-  return $string;
-});
