--- packages/twig-extensions/filters/drupal-filters/clean_id.filter.php
@@ -1,5 +1,0 @@
-<?php
-
-$filter = new Twig_SimpleFilter('clean_id', function ($string) {
-  return $string;
-});
