--- packages/twig-extensions/filters/drupal-filters/format_date.filter.php
@@ -1,5 +1,0 @@
-<?php
-
-$filter = new Twig_SimpleFilter('format_date', function ($string) {
-  return $string;
-});
