--- packages/twig-extensions/filters/drupal-filters/safe_join.filter.php
@@ -1,5 +1,0 @@
-<?php
-
-$filter = new Twig_SimpleFilter('safe_join', function ($string) {
-  return $string;
-});
