+++ packages/twig-extensions/filters/webpack.filter.php
